server/requirements.txt
fastapi
uvicorn
celery
redis
langchain_community
langchain_huggingface
python-decouple
langchain_groq
chromadb
pinecone
langchain_pinecone
sqlalchemy
databases
asyncpg
passlib
pyjwt